more code dedup, removed static utils into base class
